Let Claude extract ApplicationContext from shell.rs
